Combine legacy and liquid glass ProgressLayers
The only difference between the two was the `.glassEffect` on the liquid
glass view and the `frame`, `background` and `cornerRadius` on the
legacy view, so that can be pulled out and placed onto a new generic
progress view
